+++ src/folders.hpp
@@ -1,0 +1,6 @@
+#pragma once
+
+#include <filesystem>
+
+std::filesystem::path choose_assets_folder();
+std::filesystem::path choose_settings_folder();